pyproject.toml
[project]
name = "project-x"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "haversine>=2.9.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.6",
    "numpy>=2.3.3",
    "pandas>=2.3.2",
    "plotly-express>=0.4.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
]
